+++ plone/app/widgets/at.py
@@ -240,7 +240,8 @@ class DatetimeWidget(DateWidget):
 
         tmp = value.split(' ')
         if not tmp[0]:
-            return empty_marker
+            # empty: clear, not preserve, any existing value
+            return None, {}
         value = tmp[0].split('-')
         if len(tmp) == 2 and ':' in tmp[1]:
             value += tmp[1].split(':')